update code thêm tính năng chia sẽ liên hệ cho một nhóm, hoặc một nhân viên khác https://github.com/mynukeviet/lynx/issues/60

+++ modules/customer/site.functions.php
@@ -11,7 +11,7 @@ if (!defined('NV_MAINFILE')) die('Stop!!!');
 if (!isset($site_mods['workforce'])) {
     $workforce_list = array();
     $where = '';
-
+    
     $where .= !empty($array_config['groups_manage']) ? ' AND userid IN (SELECT userid FROM ' . NV_USERS_GLOBALTABLE . '_groups_managers WHERE group_id IN (' . $array_config['groups_manage'] . '))' : '';
     $result = $db->query('SELECT userid, first_name, last_name, username, email FROM ' . NV_USERS_GLOBALTABLE . ' WHERE active=1' . $where);
     while ($row = $result->fetch()) {
@@ -30,7 +30,7 @@ if (empty($workforce_list)) {
 function nv_crm_customer_info($customerid)
 {
     global $db;
-
+    
     $customer_info = $db->query('SELECT * FROM ' . NV_PREFIXLANG . '_customer WHERE id=' . $customerid)->fetch();
     if ($customer_info) {
         $customer_info['fullname'] = nv_show_name_user($customer_info['first_name'], $customer_info['last_name']);
@@ -41,15 +41,20 @@ function nv_crm_customer_info($customerid)
 
 function nv_customer_delete($customerid)
 {
-    global $db, $module_data;
-
-    $count = $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_customer WHERE id = ' . $customerid);
-    if ($count) {
-        // xóa dữ liệu bảng khách hàng - dịch vụ
-        //$db->exec('DELETE FROM ' . NV_PREFIXLANG . '_customer_products WHERE id_customer = ' . $customerid);
-
-        // xóa dữ liệu bảng khách hàng - sản phẩm
-        //$db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_customer_services WHERE id_customer = ' . $customerid);
+    global $db, $module_data, $user_info;
+    
+    $rows = $db->query('SELECT COUNT(*) FROM ' . NV_PREFIXLANG . '_' . $module_data . ' t1 INNER JOIN ' . NV_PREFIXLANG . '_' . $module_data . '_share_acc t2 ON t1.id=t2.customerid WHERE id=' . $customerid . ' AND t2.userid=' . $user_info['userid'] . ' AND permisson=1')->fetch();
+    if ($rows) {
+        $count = $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_customer WHERE id = ' . $customerid);
+        if ($count) {
+            // xóa dữ liệu bảng khách hàng - dịch vụ
+            //$db->exec('DELETE FROM ' . NV_PREFIXLANG . '_customer_products WHERE id_customer = ' . $customerid);
+            
+            // xóa dữ liệu bảng khách hàng - sản phẩm
+            //$db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_customer_services WHERE id_customer = ' . $customerid);
+            
+            $db->query('DELETE FROM ' . NV_PREFIXLANG . '_customer_share_acc WHERE customerid = ' . $customerid);
+        }
+        return $count;
     }
-    return $count;
 }
